Exclude pooling from manager
