+++ Assets/Unrelated Assets/Scenes/SampleScene.unity
@@ -786,6 +786,55 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 8276989356399955544, guid: 3539dbd4f14870c4c8323255190bbef1, type: 3}
   m_PrefabInstance: {fileID: 299624121}
   m_PrefabAsset: {fileID: 0}
+--- !u!1 &308047609
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 308047611}
+  - component: {fileID: 308047610}
+  m_Layer: 0
+  m_Name: ProjectilePool
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!114 &308047610
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 308047609}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 26075ffe0aae449ea4dc50f0d9739c93, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  poolSizePerType: 10
+  bulletPrefab: {fileID: 6483094454659299611, guid: 8eba6c9ab3540b247b4100d1c023fe9d, type: 3}
+  clusterPartPrefab: {fileID: 2637464847401419695, guid: 6fa1bba45d501f348b67f7ee2e9f2b3b, type: 3}
+  grenadePrefab: {fileID: 1951976267978780147, guid: 1ef7cae8c738c8a4da80c528474d9e61, type: 3}
+  rocketPrefab: {fileID: 4116445325633841098, guid: 5ee9419e7bb9f6c4bb27882fca57fbd2, type: 3}
+--- !u!4 &308047611
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 308047609}
+  serializedVersion: 2
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1001 &312701784
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -15070,11 +15119,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: c3e919d66c9f47ea995cdc8f5247ebb7, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  poolSizePerType: 10
-  bulletPrefab: {fileID: 6483094454659299611, guid: 8eba6c9ab3540b247b4100d1c023fe9d, type: 3}
-  clusterPartPrefab: {fileID: 2637464847401419695, guid: 6fa1bba45d501f348b67f7ee2e9f2b3b, type: 3}
-  grenadePrefab: {fileID: 1951976267978780147, guid: 1ef7cae8c738c8a4da80c528474d9e61, type: 3}
-  rocketPrefab: {fileID: 4116445325633841098, guid: 5ee9419e7bb9f6c4bb27882fca57fbd2, type: 3}
 --- !u!4 &2038035009
 Transform:
   m_ObjectHideFlags: 0
@@ -16120,3 +16164,4 @@ SceneRoots:
   - {fileID: 766855837}
   - {fileID: 1823771763}
   - {fileID: 2038035009}
+  - {fileID: 308047611}
